% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedata.R
\name{aggregatedata}
\alias{aggregatedata}
\title{Aggregate analysis data}
\usage{
aggregatedata(path, csvs, bins = 5, nf = NA, method = "median")
}
\arguments{
\item{path}{The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.}

\item{csvs}{--.}

\item{bins}{--.}

\item{nf}{--.}

\item{method}{--.}
}
\value{
--.
}
\description{
Aggregate analysis data
}
\examples{
\dontrun{
csvs = readcsvs (asone=TRUE)
}
}
