% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{analyze}
\alias{analyze}
\title{Analyze sounds}
\usage{
analyze(
  sound,
  from = 4800,
  to = 6800,
  nsteps = 12,
  windowlength = 0.05,
  timestep = 0.0025,
  path = NA,
  showprogress = TRUE
)
}
\arguments{
\item{sound}{a Wave object read in using the readWave function from the tuneR package, or a list containing a set of these to be analyzed. A list of paths to wave files on your hard drive can be provided instead using the path parameter.}

\item{from}{the lowest analysis frequency.}

\item{to}{the highest analysis frequency.}

\item{nsteps}{the number of steps between the lowest and highest analysis frequencies.}

\item{windowlength}{the windowlength specified in seconds.}

\item{timestep}{the analysis time step specified in seconds.}

\item{path}{a vector of file names to be analyzed, can be provided instead of the WAve objects directly.}

\item{showprogress}{if TRUE, analysis progress is shown.}
}
\value{
A list of lists of dataframes. The 'external' list is as long as number of files that were analyzed. For each 'external' list element there are N 'internal' list elements, for N analysis steps. For example, 'formant[[32]][[3]]' contains information regarding the 3rd analysis option for the 32nd file.
}
\description{
Analyze sounds
}
\examples{
\dontrun{
sound = tuneR::readWave("yoursound.wav")
ffs = analyze (sound)
plotffs (ffs)
plotffs (ffs[[2]])
}
}
