% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoselect.classic.R
\name{autoselect.classic}
\alias{autoselect.classic}
\title{Automatically select winners}
\usage{
autoselect.classic(
  formants,
  order = 5,
  n_formants = 4,
  outputpath = NA,
  subset = NA,
  progressbar = FALSE
)
}
\arguments{
\item{formants}{a list of formant data read in with the readformants function.}

\item{order}{the order of the prediction model.}

\item{n_formants}{the number of formants to optimize for.}

\item{outputpath}{if NA, nothing is written out. If "working", data is written out to the working directory. Any other path may also be specified.}

\item{subset}{a vector indicating a subset of the analyses to be considered.}

\item{progressbar}{if TRUE, a progress bar prints out in the console.}
}
\value{
A list containing information about the selection of the winners. The list elements are:

1) winners_csv: A dataframe containing the same information as the winners.csv file. 

2) errors: A 3d array containing information about the RMS error for each analysis for each formant. Dimensions are [file, analysis, formant]. 

3) total_errors: A 2d array containing information about the total RMS error for each analysis. Dimensions are [file, analysis]. 
4) coefficients: A 4d array containing information about regression coefficients for prediction of each formant for each analysis. Dimensions are [file, analysis, formant, coefficient]. Coefficients are arranged in terms of increasing order (i.e., intercept, linear term, quadratric term,...).
}
\description{
Select the best analyses using the 'classic' Fast Track method from Praat. Generates identical output to the autoselect step in Praat, except for no regression information text files are written (for now).
}
\examples{
\dontrun{
formants = readformants ()
winners = autoselect.classic (formants, progressbar = TRUE)
winners = autoselect.classic (formants, outputpath="working")
}
}
