% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoselectwinners.R
\name{autoselectwinners}
\alias{autoselectwinners}
\title{Automatically select winning analyses}
\usage{
autoselectwinners(
  formants,
  order = 5,
  n_formants = 4,
  method = "classic",
  outputpath = NA,
  subset = NA
)
}
\arguments{
\item{formants}{a list of formant data read in with the readformants function.}

\item{order}{the order of the prediction model.}

\item{n_formants}{the number of formants to optimize for.}

\item{method}{method of selecting the winning analysis.}

\item{outputpath}{--.}

\item{subset}{a vector indicating a subset of the analyses to be considered.}
}
\value{
A vector with the winning analysis for each file.
}
\description{
Automatically select winning analyses
}
\examples{
\dontrun{
formants = readformants ()
winners = autoselectwinners (formants, outputpath="working")
}
}
