% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefficients}
\alias{coefficients}
\title{Example of Fat Track regressions coefficients}
\format{
An object of class \code{data.frame} with 72 rows and 19 columns.
}
\source{
\url{https://github.com/santiagobarreda/FastTrackR}
}
\usage{
coefficients
}
\description{
An example of the analysis regression coefficients provided by Fast Track (in an 'coefficients.csv' file). The data is productions of 12 vowels by 7 adult males from the Hillenbrand et al. (1995) data. The coefficients columns are so that column cXY represents the coefficient for the Xth formant, for the Yth time slice.
}
\details{
Data representing productions of 12 vowels by x adult males from the
Hillenbrand et al. (1995) data.
}
\section{Columns}{

\describe{
\item{file}{the recording file name.}
\item{c11}{The 0th order coefficient (intercept) for the for the first formant.}
\item{c21}{A series of columns, each representing the Yth order coefficients for formant X.}
\item{...}{And so on.}
}
}

\references{
Hillenbrand, J.M., Getty, L.A., Clark, M.J., and Wheeler, K. (1995). "Acoustic characteristics of American English vowels," Journal of the Acoustical Society of America, 97, 3099-3111.
}
\keyword{datasets}
