% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{contours}
\alias{contours}
\title{Plot formant contours from aggregated files}
\usage{
contours(path)
}
\arguments{
\item{path}{The path to your Fast Track working directory.}
}
\value{
A list containing the formant objects as data frames.
}
\description{
Plot formant contours from aggregated files
}
\examples{
1+1
}
