% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{downsample}
\alias{downsample}
\title{Downsample}
\usage{
downsample(sound, maxformant = 5000, precision = 50)
}
\arguments{
\item{sound}{a numeric vector representing the sound to be analyzed.}

\item{maxformant}{the desired maximum analysis frequency (i.e., the new Nyquist/2).}

\item{precision}{the number of neighbors used to interpolate.}
}
\value{
A numeric vector representing the downsampled sound. The new sampling frequency is maxformant*2.
}
\description{
Downsample
}
\examples{
\dontrun{
sound = tuneR::readWave("yoursound.wav")
tmp_snd = downsample (snd, maxformant = 5000)
}
}
