% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrids.R
\name{extract.internal}
\alias{extract.internal}
\title{Extract Vowels}
\usage{
\method{extract}{internal}(
  tgpath,
  sndpath,
  segmenttier = "phone",
  wordtier = NA,
  commenttiers = NA,
  omittier = NA,
  stress = c(0, 1, 2),
  wordstoskip = NA
)
}
\arguments{
\item{tgpath}{a path or vector of paths to textgrid files.}

\item{sndpath}{a path or vector of paths to wav files.}

\item{segmenttier}{the name of the tier containing segmental information used for extraction.}

\item{wordtier}{the name of a tier containing information about words (optional).}

\item{commenttiers}{a vector containing names of tiers with comments you wish to extract (optional).}

\item{omittier}{the name of a tier indicating which segments you wish to skip (optional).}

\item{stress}{a vector contianing labels you may have used to mark stress.}

\item{wordstoskip}{a vector containing any words you do not want to extract vowels from.}
}
\description{
Extracts vowels from larger sounds using information from matching TextGrids. A complete explanation of function behavior is provided in the extractvowels reference page.
}
\examples{
\dontrun{
tgpath = "yoursound2.TextGrid"
tmp = readtextgrid (tgpath)
output = extract.internal (tgpath, wordtier="word",commenttiers=c("omit","comments"))
}
}
