% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrids.R
\name{extractsounds}
\alias{extractsounds}
\title{Extract Sounds}
\usage{
extractsounds(
  tgpath,
  sndpath,
  outputpath = "output",
  segmenttier = "phone",
  wordtier = NA,
  commenttiers = NA,
  omittier = NA,
  stress = c(0, 1, 2),
  wordstoskip = NA
)
}
\arguments{
\item{tgpath}{--.}

\item{sndpath}{--.}

\item{outputpath}{--.}

\item{segmenttier}{--.}

\item{wordtier}{--.}

\item{commenttiers}{--.}

\item{omittier}{--.}

\item{stress}{--.}

\item{wordstoskip}{--.}
}
\value{
--.
}
\description{
Extracts vowels from larger sounds using information from matching TextGrids.Words to skip does not work, comment tiers do not work.
}
\examples{
\dontrun{
# nothing yet
}
}
