% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ft.arrows}
\alias{ft.arrows}
\title{Plot arrows}
\usage{
ft.arrows(
  aggregated_data,
  xformant = 1,
  yformant = 2,
  revaxes = FALSE,
  logaxes = FALSE,
  lwd = 2,
  length = 0.1,
  add = TRUE,
  ...
)
}
\arguments{
\item{aggregated_data}{A dataframe containing the data from an 'aggregated_data' file produced by Fast Track.}

\item{xformant}{an integer indicating which formant number should be plotted on the x axis.}

\item{yformant}{an integer indicating which formant number should be plotted on the y axis.}

\item{revaxes}{if TRUE, axis ranges are inverted so that they go high > low.}

\item{logaxes}{if TRUE, axes are logarithmic.}

\item{lwd}{an integer determining arrow line width.}

\item{length}{the length of the arrow lines.}

\item{add}{if FALSE, a new plot if created.}

\item{...}{Additional arguments are passed to the internal call of 'arrows'.}
}
\description{
Add arrows to the end of contours made by the ft.contour function. Colors are taken from the 'color' column in the data.
}
\examples{
data(aggregated_data)
ft.lines (aggregated_data, xformant=2,yformant=1,revaxes=TRUE)
ft.arrows (aggregated_data, xformant=2,yformant=1,revaxes=TRUE)
}
