% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ft.lines}
\alias{ft.lines}
\title{Plot formant contours}
\usage{
ft.lines(
  aggregated_data,
  xformant = 1,
  yformant = 2,
  revaxes = FALSE,
  logaxes = FALSE,
  lwd = 2,
  starttime = 0,
  endtime = 0,
  add = FALSE,
  ...
)
}
\arguments{
\item{aggregated_data}{A dataframe containing the data from an 'aggregated_data' file produced by Fast Track.}

\item{xformant}{an integer indicating which formant number should be plotted on the x axis.}

\item{yformant}{an integer indicating which formant number should be plotted on the y axis.}

\item{revaxes}{if TRUE, axis ranges are inverted so that they go high > low.}

\item{logaxes}{if TRUE, axes are logarithmic.}

\item{lwd}{an integer determining contour line width.}

\item{starttime}{an integer indicating which time point the contour should start at.}

\item{endtime}{an integer indicating which time point the contour should end at.}

\item{add}{if FALSE, a new plot if created.}

\item{...}{Additional arguments are passed to the internal call of 'lines'.}
}
\description{
Draws lines representing formant contours using the information represented in aggregated data files. Colors are taken from the 'color' column in the data.
}
\examples{
data(aggregated_data)
ft.lines (aggregated_data)
ft.lines (aggregated_data, xformant=2,yformant=1,revaxes=TRUE)
}
