% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ft.plot}
\alias{ft.plot}
\title{Plot an aggregated file}
\usage{
ft.plot(
  aggregated_data,
  xformant = 1,
  yformant = 2,
  revaxes = FALSE,
  logaxes = FALSE,
  xlab,
  ylab,
  add = FALSE,
  ...
)
}
\arguments{
\item{aggregated_data}{A dataframe containing the data from an 'aggregated_data' file produced by Fast Track.}

\item{xformant}{an integer indicating which formant number should be plotted on the x axis.}

\item{yformant}{an integer indicating which formant number should be plotted on the y axis.}

\item{revaxes}{if TRUE, axis ranges are inverted so that they go high > low.}

\item{logaxes}{if TRUE, axes are logarithmic.}

\item{xlab}{an optional user-specified x-axis label.}

\item{ylab}{an optional user-specified y-axis label.}

\item{add}{if FALSE, a new plot if created.}

\item{...}{Additional arguments are passed to the internal call of 'plot'.}
}
\description{
Set up a plot aggregated files. This function does not draw anything but is used to set up the plot for other functions.
}
