% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ft.points}
\alias{ft.points}
\title{Plot formant contours from aggregated files}
\usage{
ft.points(
  aggregated_data,
  xformant = 1,
  yformant = 2,
  revaxes = FALSE,
  logaxes = FALSE,
  cex = 2,
  time = 1,
  add = TRUE,
  ...
)
}
\arguments{
\item{aggregated_data}{A dataframe containing the data from an 'aggregated_data' file produced by Fast Track.}

\item{xformant}{an integer indicating which formant number should be plotted on the x axis.}

\item{yformant}{an integer indicating which formant number should be plotted on the y axis.}

\item{revaxes}{if TRUE, axis ranges are inverted so that they go high > low.}

\item{logaxes}{if TRUE, axes are logarithmic.}

\item{cex}{an integer determining point size.}

\item{time}{an integer indicating which time point the point should be placed at.}

\item{add}{if FALSE, a new plot if created.}

\item{...}{Additional arguments are passed to the internal call of 'text'.}
}
\description{
Plot formant contours from aggregated files
}
\examples{
data(aggregated_data)
ft.contours (aggregated_data, xformant=2,yformant=1,revaxes=TRUE)
ft.arrows (aggregated_data, xformant=2,yformant=1,revaxes=TRUE)
ft.points (aggregated_data, xformant=2,yformant=1,revaxes=TRUE)
}
