% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwinners.R
\name{getwinners}
\alias{getwinners}
\title{Get winning analyses}
\usage{
getwinners(path, winners, formants, asone = TRUE, write = FALSE)
}
\arguments{
\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}

\item{winners}{an optional dataframe representing the data in the 'winners.csv' file.}

\item{formants}{an optional list of lists representing all of the possible formant analyses.}

\item{asone}{if TRUE, the formant data is stuck together into one dataframe and filenames are indicated in a new column. If FALSE, a list of dataframes is returned and each list element is named after the file.}

\item{write}{if TRUE, the data is written out to a CSV folder in the working directory.}
}
\value{
A dataframe or list of dataframes, as per the asone parameter.
}
\description{
This function will read winning analyses into R based on the content of the 'winners.csv' file. The required formant and csv data can be pre-loaded with the readformants() and readcsvs() functions provided in this package. Reading the formant data can be a bit slow so loading it once and passing it to the function will speed things up quite a bit if this function will be called repeatedly.
}
\examples{
\dontrun{
getwinners ()
}
}
