% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposebounds.R
\name{imposebounds}
\alias{imposebounds}
\title{Impose boundaries on formants}
\usage{
imposebounds(path, boundaries = NA, selectioninfo = NA, write = TRUE)
}
\arguments{
\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}

\item{boundaries}{a dataframe representing the boundaries for your formants for each vowel.}

\item{selectioninfo}{--.}

\item{write}{if TRUE, a new winners file will be written to your folder.}
}
\value{
A dataframe or list of dataframes, as per the asone parameter.
}
\description{
Create an updated winners dataframe or file with imposed formant boundaries.
}
\examples{
\dontrun{
 # coming soon
}
}
