% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeheuristics.R
\name{imposeheuristics}
\alias{imposeheuristics}
\title{Impose boundaries on formants}
\usage{
imposeheuristics(
  path = NA,
  f1max = 1200,
  f4min = 3800,
  rhotic = c(2000, 500),
  minF3F4 = c(500, 1500),
  selectioninfo = NA,
  write = FALSE
)
}
\arguments{
\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}

\item{f1max}{--.}

\item{f4min}{--.}

\item{rhotic}{--.}

\item{minF3F4}{--.}

\item{selectioninfo}{--.}

\item{write}{if TRUE, a new winners file will be written to your folder, reflecting the penalties.}
}
\value{
A dataframe or list of dataframes, as per the asone parameter.
}
\description{
Create an updated winners dataframe or file with imposed penalties on formants. Penalties can be determined with the imposeboundaries or imposeheuristics functions included in this package.
}
\examples{
\dontrun{
 # coming soon
}
}
