% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadformants.R
\name{loadformants}
\alias{loadformants}
\title{Load Praat formant objects}
\usage{
loadformants(path)
}
\arguments{
\item{path}{The path to your Fast Track working directory.}
}
\value{
A list containing the formnt objects as data frames.
}
\description{
Load Praat formant objects
}
\examples{
1+1
}
