% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecomparisonplots.R
\name{makecomparisonplots}
\alias{makecomparisonplots}
\title{Make images comparing alternative analysies}
\usage{
makecomparisonplots(
  formants_plot,
  path = NA,
  height = 1000,
  width = 1400,
  pointsize = 20,
  winners = NA,
  spectrogram_timestep = 0.0015,
  showprogressbar = TRUE,
  ...
)
}
\arguments{
\item{formants_plot}{The formant data to be used for making plots. Read in using the readformants function in this package.}

\item{path}{The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.}

\item{height}{height of each plot in pixels.}

\item{width}{width of each plot in pixels.}

\item{pointsize}{point size for plotting.}

\item{winners}{.}

\item{spectrogram_timestep}{--.}

\item{showprogressbar}{--.}

\item{...}{--.}
}
\description{
Make images comparing alternative analysies
}
\examples{
\dontrun{
makecomparisonplots(formants[1:10])
makecomparisonplots(formants[1:10], winners = winners$winners_csv[1:10,])
}
}
