% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeimages.R
\name{makewinnerplots}
\alias{makewinnerplots}
\title{MAke images for winning analysies}
\usage{
makewinnerplots(path, resolution = 1000, csvs = NA)
}
\arguments{
\item{path}{--.}

\item{resolution}{--.}

\item{csvs}{--.}
}
\description{
MAke images for winning analysies
}
\examples{
\dontrun{
# nothing yet
}
}
