% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.lan.R
\name{ourliers.lan}
\alias{ourliers.lan}
\title{Find outliers using Log-Additive Normalization}
\usage{
ourliers.lan(aggregated_data)
}
\arguments{
\item{aggregated_data}{--.}
}
\value{
--.
}
\description{
Find outliers using Log-Additive Normalization
}
\examples{
\dontrun{
 # coming soon
}
}
