% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.lda.R
\name{ourliers.lda}
\alias{ourliers.lda}
\title{Find outliers using Linear Discriminant Analysis}
\usage{
ourliers.lda(aggregated_data)
}
\arguments{
\item{aggregated_data}{--.}
}
\value{
--.
}
\description{
Find outliers using Linear Discriminant Analysis
}
\examples{
\dontrun{
 # coming soon
}
}
