% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.lda.R
\name{outliers.lda}
\alias{outliers.lda}
\title{Find outliers using Linear Discriminant Analysis}
\usage{
outliers.lda(aggregated_data)
}
\arguments{
\item{aggregated_data}{--.}
}
\value{
--.
}
\description{
This is mostly a placeholder.
}
\examples{
\dontrun{
 # coming soon
}
}
