% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalizewinners.R
\name{penalizewinners}
\alias{penalizewinners}
\title{Impose boundaries on formants}
\usage{
penalizewinners(path, selectioninfo = NA, write = FALSE)
}
\arguments{
\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}

\item{selectioninfo}{--.}

\item{write}{if TRUE, a new winners file will be written to your folder, reflecting the penalties.}
}
\value{
A dataframe or list of dataframes, as per the asone parameter.
}
\description{
Create an updated winners dataframe or file with imposed penalties on formants. Penalties can be determined with the imposeboundaries or imposeheuristics functions included in this package.
}
\examples{
\dontrun{
 # coming soon
}
}
