% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readformants.R
\name{readformants}
\alias{readformants}
\title{Load Fast Track formant objects}
\usage{
readformants(
  path,
  fileinformation = NA,
  progressbar = TRUE,
  encoding = "UTF-8"
)
}
\arguments{
\item{path}{The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.}

\item{fileinformation}{a dataframe representing the "file_information.csv" file used by Fast Track. If NA, it is loaded from the working directory.}

\item{progressbar}{if TRUE, a progress bar prints out in the console.}

\item{encoding}{--.}
}
\value{
A list of lists of dataframes. The "external" list is as long as number of files that were analyzed. For each "external" list element there are \emph{n} "internal" list elements for \emph{n} analysis steps. For example, \code{formant[[32]][[3]]} contains information regarding the 3rd analysis option for the 32nd file.
}
\description{
This function reads in all the files contained in the "/formants" folder within a Fast Track directory. Since Fast Track exports one formant object per analysis option per token, there are usually a large number of files stored in that directory. This function makes it easy to read all that information in at once. Note that it may take a bit of time (1-2 minutes per thousand sound files) for all the data to be read in if there are many tokens that were analyzed in that directory.
}
\details{
This function returns objects of the class 'formants' that contain all information about all analyses for all files, in addition to some other information. 'formants' objects are lists of lists of dataframes. At the top level is the list containing all information for all files. This list has n elements for n files. Each list element is a list of x dataframe, for an analysis with x different analysis steps. Each individual analysis is stored in a  data frame containing formant frequencies and bandwidths for the first 3-4 formants. These values are rounded, which makes display easier and makes the resulting object much smaller in memory (and on your hard drive). 

So, if you read in your analysis into an object called 'formants', then 'formants[[2]][[3]]' represents the dataframe containing the third analysis for the second sound file. 'formants[[2]]' is a list containing all of the dataframes for the second sound file and 'formants' is a list containing each of the list of dataframes for all the sound files. 

Every data frame representing a single Praat formant object needs to have some information associated with it, and this is done using attributes for the object. The attributes for every dataframe representing a formant object include: timestep (the analysis time step, in ms), w1 (the location of the first analysis window, in ms), maxformant (the maximum formant frequency, in Hz), and filename, containing the label used for the formant object (relating to the wav file filename). 

Both the toplevel list and the lists interior to the 'formants' object have attributes that contain information that might be useful to the user, such as: analysis cutoffs, sound labels, number of files, number of analysis levels, time step, duration, and more.
}
\examples{
\dontrun{
formants = readformants (progressbar = TRUE)
}
}
