% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsounds.R
\name{readsounds}
\alias{readsounds}
\title{Read sounds a single data file}
\usage{
readsounds(path = NA)
}
\arguments{
\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}
}
\description{
This function writes out R internal formant data into text files that can be used in Praat.
}
