% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtextgrid.R
\name{readtextgrid}
\alias{readtextgrid}
\title{Load textgrid information into R}
\usage{
readtextgrid(path, encoding = "UTF-8")
}
\arguments{
\item{path}{the path to the Praat textgrid file you want to read.}

\item{encoding}{--.}
}
\value{
A list of dataframes, one containing the data from each interval tier in the textgrid. Each list element is named after the tier.
}
\description{
Creates a list of dataframes from a textgrid. Each list element is a dataframe representing information from a different interval tier. Each dataframe contains the interval label, start time, end time, and duration (all in milliseconds).
}
\examples{
\dontrun{
path = "tuvesamisol.TextGrid"

path = "testing.txt"
path = "WS005-Melissa.TextGrid"
readtextgrid (path)
}
}
