% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwinners.R
\name{readwinners}
\alias{readwinners}
\title{Load information about the winning analyses}
\usage{
readwinners(path = NA)
}
\arguments{
\item{path}{The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.}
}
\value{
A dataframe containing information about winning analyses.
}
\description{
This function reads in the winners.csv file from a specified path, or assumes it is located in the working directory if not path is specified.
}
\examples{
\dontrun{
winners <- readwinners()
}
}
