% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{textgrid_example}
\alias{textgrid_example}
\title{Example of a Praat TextGrid File}
\format{
An object of class \code{character} of length 604.
}
\usage{
textgrid_example
}
\description{
Example of a Praat TextGrid File
}
\keyword{datasets}
