% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackformants.folder.R
\name{trackformants.folder}
\alias{trackformants.folder}
\title{Track formants for a folder of sounds}
\usage{
trackformants.folder(
  path = NA,
  from = 4800,
  to = 6800,
  nsteps = 12,
  windowlength = 0.05,
  write = TRUE,
  n_formants = 3,
  timestep = 0.002,
  fileinformation = NA,
  progressbar = TRUE,
  write_infos = FALSE,
  sounds = NA,
  encoding = "UTF-8"
)
}
\arguments{
\item{path}{the path to the folder you with to analyze. If NA, the working directory is used.}

\item{from}{the lowest maximum formant frequency that will be considered.}

\item{to}{the highest maximum formant frequency that will be considered.}

\item{nsteps}{the number of analysis steps between the lowest and highest analysis frequencies.}

\item{windowlength}{the LPC analysis window length specified in seconds.}

\item{write}{if TRUE, the result of the analysis is saved as an RDS file.}

\item{n_formants}{the number of formants returned and optimized over (5.5. are always tracked).}

\item{timestep}{the analysis time step specified in seconds.}

\item{fileinformation}{a dataframe representing the "file_information.csv" file used by Fast Track. If NA, it is loaded from the working directory.}

\item{progressbar}{if TRUE, information about estimated analysis time is printed.}

\item{write_infos}{if TRUE, analysis info text files are written out as in Praat.}

\item{sounds}{a list of previously loaded sounds.}

\item{encoding}{if loading file_information from a local CSV file, you may need to specify the encoding or change the file to UTF-8 encoding.}
}
\value{
A 'formants' object containing information about all alternate formant analyses for all files. See the help for the readformants() function in this package for more information on 'formants' objects.
}
\description{
Track formants for a folder of sounds
}
\examples{
\dontrun{
ffs = trackformants.folder ()
ffs
plotffs (ffs)
plotffs (ffs[[2]])
}
}
