% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackformants.R
\name{trackformants}
\alias{trackformants}
\title{Track formants for a sound or folder of sounds}
\usage{
trackformants(
  path = NA,
  from = 4800,
  to = 6800,
  nsteps = 12,
  windowlength = 0.05,
  write = TRUE,
  n_formants = 3,
  timestep = 0.002,
  fileinformation = NA,
  progressbar = TRUE,
  write_infos = FALSE
)
}
\arguments{
\item{path}{a Wave object read in using the readWave function from the tuneR package, or a list containing a set of these to be analyzed. A list of paths to wave files on your hard drive can be provided instead using the path parameter.}

\item{from}{the lowest analysis frequency.}

\item{to}{the highest analysis frequency.}

\item{nsteps}{the number of steps between the lowest and highest analysis frequencies.}

\item{windowlength}{the windowlength specified in seconds.}

\item{write}{if TRUE, the result of the analysis is saved as an RDS file.}

\item{n_formants}{the number of formants returned (5.5. are always tracked).}

\item{timestep}{the analysis time step specified in seconds.}

\item{fileinformation}{a dataframe representing the "file_information.csv" file used by Fast Track. If NA, it is loaded from the working directory.}

\item{progressbar}{if TRUE, information about estimated analysis time is printed.}

\item{write_infos}{if TRUE, analysis info text files are written out as in Praat.}
}
\value{
A dataframe containing information about the formant tracks for the file.
}
\description{
Track formants for a sound or folder of sounds
}
\examples{
\dontrun{
sound = readWave2("yoursound.wav")
ffs = trackformants (sound, timestep = 0.002)
plotffs (ffs)
plotffs (ffs[[2]])
}
}
