% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatewinners.R
\name{updatewinners}
\alias{updatewinners}
\title{Update winners after changing penalties}
\usage{
updatewinners(path = NA, selection_info = NA, write = FALSE)
}
\arguments{
\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}

\item{selection_info}{--.}

\item{write}{if TRUE, ....}
}
\value{
A dataframe or list of dataframes, as per the asone parameter.
}
\description{
Update winners after changing penalties
}
\examples{
\dontrun{
getwinners ()
tmp = getwinners (getwd(), winners$winners, formants, write = TRUE)
}
}
