% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels}
\alias{vowels}
\title{Vowel plotting label information}
\format{
An object of class \code{data.frame} with 30 rows and 4 columns.
}
\usage{
vowels
}
\description{
A dataframe containing ARPABET and IPA information for vowel symbols.
}
\section{Columns}{

\describe{
\item{IPA}{the IPA symbol for each vowel. }
\item{arpabet}{the ARPABET representation for each vowel. }
\item{unicode}{the unicode representation for each vowel. }
\item{description}{A verbal description of the vowel sound. }
}
}

\keyword{datasets}
