% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowelstoextract}
\alias{vowelstoextract}
\title{Default set of vowel symbols to extract from Praat TextGrids}
\format{
An object of class \code{data.frame} with 45 rows and 3 columns.
}
\source{
\url{https://github.com/santiagobarreda/FastTrackR}
}
\usage{
vowelstoextract
}
\description{
A default set of vowel symbols to extract from Praat TextGrids.
}
\section{Columns}{

\describe{
\item{group}{group number is useless i nthis context, I think, but is maintained to keep the same formatting as in the Praat file. }
\item{label}{the labels used to indicate vowels that should be extracted.}
\item{color}{which color will be used when using some automatic plotting functions. Also more useful in Praat but maintained for consistency.}
}
}

\keyword{datasets}
