% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wav_example}
\alias{wav_example}
\title{Example of a wav file
#'}
\format{
An object of class \code{Wave} of length 162202.
}
\usage{
wav_example
}
\description{
Example of a wav file
#'
}
\keyword{datasets}
