% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writecsvs.R
\name{writecsvs}
\alias{writecsvs}
\title{Write Fast Track csv files}
\usage{
writecsvs(csvs, path)
}
\arguments{
\item{csvs}{csv data read in using the readcsvs() function, either as a dataframe or a list of dataframes.}

\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}
}
\description{
Write internal formant data to csv files.
}
\examples{
\dontrun{
csvs = readcsvs ()
}
}
