% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{writeformants}
\alias{writeformants}
\title{Write Fast Track formant objects}
\usage{
writeformants(formants, path)
}
\arguments{
\item{formants}{--.}

\item{path}{--.}
}
\description{
Write Fast Track formant objects
}
