% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writesounds.R
\name{writesounds}
\alias{writesounds}
\title{Write sound objects}
\usage{
writesounds(sounds, path = NA)
}
\arguments{
\item{sounds}{a list of sounds read in with the readsounds function of extracted from a textgrid with the extractvowels function.}

\item{path}{the path to the working directory for the Fast Track project. If no path is provided this is the working directory.}
}
\description{
This function writes sound objects as a single data file or as a set of WAV files.
}
