% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoselect.classic.R
\name{autoselect.classic}
\alias{autoselect.classic}
\title{Automatically select winners}
\usage{
autoselect.classic(
  formants,
  order = 5,
  n_formants = 4,
  outputpath = NA,
  subset = NA,
  progressbar = FALSE
)
}
\arguments{
\item{formants}{a list of formant data read in with the readformants function.}

\item{order}{the order of the prediction model.}

\item{n_formants}{the number of formants to optimize for.}

\item{outputpath}{--.}

\item{subset}{a vector indicating a subset of the analyses to be considered.}

\item{progressbar}{if TRUE, a progress bar prints out in the console.}
}
\value{
A vector with the winning analysis for each file.
}
\description{
Select the best analyses using the 'classic' Fast Track method from Praat. Generates identical output to the autoselect step in Praat, except for no regression information text files are written (for now).
}
\examples{
\dontrun{
formants = readformants ()
winners = autoselect.classic (formants)
winners = autoselect.classic (formants, outputpath="working")
}
}
