% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoselect.classic.R
\name{autoselect.classic}
\alias{autoselect.classic}
\title{Automatically select winners}
\usage{
autoselect.classic(
  formants,
  order = 5,
  n_formants = 4,
  outputpath = NA,
  subset = NA,
  progressbar = FALSE,
  write = TRUE
)
}
\arguments{
\item{formants}{a list of formant data read in with the readformants function, or created using the trackformants function.}

\item{order}{the order of the prediction model.}

\item{n_formants}{the number of formants to optimize for.}

\item{outputpath}{if NA, nothing is written out. If "working", data is written out to the working directory. Any other path may also be specified.}

\item{subset}{a vector indicating a subset of the analyses to be considered.}

\item{progressbar}{if TRUE, a progress bar prints out in the console.}

\item{write}{if TRUE, an RDS file containing selection information is saved to the working directory.}
}
\value{
An object of the class "selection_info". A list containing information about the selection of the winners. See the documentation for readselectioninfo for more information.
}
\description{
Select the best analyses using the 'classic' Fast Track method from Praat. Each formant is predicted using a regression model, and the analysis with the smallest . Generates identical output to the autoselect step in Praat, except for no regression information text files are written (for now).
}
\examples{
\dontrun{
# load a previous analysis from Praat
formants = readformants ()

# or load a previous analysis saved from R
formants = readRDS ('formants.RDS')

# or track the formants using R
formants = trackformants ()

# keep results in R
winners = autoselect.classic (formants, progressbar = TRUE)

# generate Praat compliant data files
winners = autoselect.classic (formants, outputpath="working")
}
}
