% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrids.R
\name{extract.internal}
\alias{extract.internal}
\title{Extract Vowels}
\usage{
extract.internal(
  tgpath,
  sndpath,
  segmenttier = "phone",
  wordtier = NA,
  commenttiers = NA,
  omittier = NA,
  stress = c(0, 1, 2),
  wordstoskip = NA
)
}
\arguments{
\item{tgpath}{--.}

\item{sndpath}{--.}

\item{segmenttier}{--.}

\item{wordtier}{--.}

\item{commenttiers}{--.}

\item{omittier}{--.}

\item{stress}{--.}

\item{wordstoskip}{--.}
}
\value{
--.
}
\description{
Extracts vowels from larger sounds using information from matching TextGrids.
}
\examples{
\dontrun{
q = readWave2 ("yoursound2.wav")
tgpath = "yoursound2.TextGrid"
tmp = readtextgrid (tgpath)
output = extractsounds (tgpath, wordtier="word",commenttiers=c("omit","comments"))
}
}
