% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractvowels.R
\name{extractvowels}
\alias{extractvowels}
\title{Extract Vowels}
\usage{
extractvowels(
  tgpath = NA,
  sndpath = NA,
  outputpath = NA,
  segmenttier = 1,
  wordtier = NA,
  commenttiers = NA,
  omittier = NA,
  stress = c(0, 1, 2),
  wordstoskip = NA
)
}
\arguments{
\item{tgpath}{a path or vector of paths to textgrid files.}

\item{sndpath}{a path or vector of paths to wav files.}

\item{outputpath}{a path to the folder where you want data to go. If NA, nothing is written out. If "working" it will output in a folder called "output" in your working directory.}

\item{segmenttier}{the name of the tier containing segmental information used for extraction.}

\item{wordtier}{the name of a tier containing information about words (optional).}

\item{commenttiers}{a vector containing names of tiers with comments you wish to extract (optional).}

\item{omittier}{the name of a tier indicating which segments you wish to skip (optional).}

\item{stress}{a vector contianing labels you may have used to mark stress.}

\item{wordstoskip}{a vector containing any words you do not want to extract vowels from.}
}
\value{
--.
}
\description{
Extracts vowels from larger sounds using information from matching TextGrids.
}
\examples{
\dontrun{
tgpath = "tuvesamisol.TextGrid"
tgpath = "WS005-Melissa.TextGrid"

output = extractvowels (tgpath, stress = NA)
output = extractvowels (tgpath, wordtier=NA)
output = extractvowels (tgpath, wordtier=2)
}

}
