% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findformants.R
\name{findformants}
\alias{findformants}
\title{Carry out LPC for sound}
\usage{
findformants(
  sound,
  n_formants = 4,
  maxformant = 7000,
  windowlength = 0.05,
  timestep = 0.002,
  preemphasis_frequency = 50,
  label = NA,
  returnsound = FALSE
)
}
\arguments{
\item{sound}{a numeric vector representing the sound to be analyzed.}

\item{n_formants}{--.}

\item{maxformant}{the maximum analysis frequency (i.e., the Nyquist/2).}

\item{windowlength}{the windowlength specified in seconds.}

\item{timestep}{the analysis time step specified in seconds.}

\item{preemphasis_frequency}{--.}

\item{label}{--.}

\item{returnsound}{--.}
}
\value{
A matrix in which each row represents a different time point. The first four columns represent the frequencies of F1-F4, and columns 5-8 represent their bandwidths.
}
\description{
Carry out LPC for sound
}
\examples{
\dontrun{
sound = readWave2("yoursound.wav")
trackformants (sound, maxformant = 5300)
sound = readWave2("C:/Users/santi/Desktop/JSdata/melissa-R/sounds/WS005-Melissa_2725.wav")
sound = readWave2("C:/Users/santi/Desktop/JSdata/melissa-R/sounds/WS005-Melissa_0001_14000.wav")
}
}
