% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecomparisonplots.R
\name{makecomparisonplots}
\alias{makecomparisonplots}
\title{Make images comparing alternative analysies}
\usage{
makecomparisonplots(
  formants_plot,
  path = NA,
  height = 1000,
  width = 1400,
  pointsize = 20,
  winners = NA,
  number_of_lines = 0.0015,
  progressbar = TRUE,
  alternate_output_path = NA,
  ...
)
}
\arguments{
\item{formants_plot}{The formant data to be used for making plots. Read in using the readformants function in this package.}

\item{path}{The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.}

\item{height}{height of each plot in pixels.}

\item{width}{width of each plot in pixels.}

\item{pointsize}{point size for plotting.}

\item{winners}{if you want winners to be highlighted, provide a dataframe with the information in the winners.csv file.}

\item{number_of_lines}{the number of pixels along the x axis for each spectrogram. Consider in relation to image width.}

\item{progressbar}{if TRUE, information about estimated analysis time is printed.}

\item{alternate_output_path}{if not NA, images will be written directly to this path.}

\item{...}{additional parameters are passed to the internal call of plotffs.}
}
\description{
Make images comparing alternative analysies
}
\examples{
\dontrun{
makecomparisonplots(formants[1:10])
makecomparisonplots(formants[1:10], winners = winners$winners_csv[1:10,])
}
}
