% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecomparisonplots.R
\name{makecomparisonplots}
\alias{makecomparisonplots}
\title{Make images comparing alternative analysies}
\usage{
makecomparisonplots(
  path,
  resolution = 1000,
  formants = NA,
  winners = NA,
  showprogressbar = FALSE
)
}
\arguments{
\item{path}{--.}

\item{resolution}{--.}

\item{formants}{--.}

\item{winners}{--.}

\item{showprogressbar}{--.}
}
\description{
Make images comparing alternative analysies
}
\examples{
\dontrun{
system.time (makecomparisonplots(showprogressbar=TRUE))
}
}
