% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecomparisonplots.R
\name{makecomparisonplots}
\alias{makecomparisonplots}
\title{Make images comparing alternative analysies}
\usage{
makecomparisonplots(
  formants_plot,
  path = NA,
  height = 1000,
  width = 1400,
  pointsize = 20,
  winners = NA,
  spectrogram_timestep = 0.0015,
  showprogressbar = TRUE,
  ...
)
}
\arguments{
\item{formants_plot}{--.}

\item{path}{--.}

\item{height}{--.}

\item{width}{--.}

\item{pointsize}{--.}

\item{winners}{--.}

\item{spectrogram_timestep}{--.}

\item{showprogressbar}{--.}

\item{...}{--.}
}
\description{
Make images comparing alternative analysies
}
\examples{
\dontrun{
makecomparisonplots(formants[1:10])
makecomparisonplots(formants[1:10], winners = winners$winners_csv[1:10,])
}
}
