% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeimages.R
\name{makecomparisonplots}
\alias{makecomparisonplots}
\title{Make images comparing alternative analysies}
\usage{
makecomparisonplots(path, resolution = 1000, formants = NA, winners = NA)
}
\arguments{
\item{path}{--.}

\item{resolution}{--.}

\item{formants}{--.}

\item{winners}{--.}
}
\description{
Make images comparing alternative analysies
}
\examples{
\dontrun{
# nothing yet
}
}
