% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makewinnerplots.R
\name{makewinnerplots}
\alias{makewinnerplots}
\title{Make images for winning analysies}
\usage{
makewinnerplots(
  path = NA,
  csvs = NA,
  height = 1000,
  width = 1400,
  pointsize = 20,
  number_of_lines = 250,
  showprogressbar = TRUE
)
}
\arguments{
\item{path}{The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.}

\item{csvs}{.}

\item{height}{--.}

\item{width}{--.}

\item{pointsize}{--.}

\item{number_of_lines}{--.}

\item{showprogressbar}{--.}
}
\description{
Make images for winning analysies
}
\examples{
\dontrun{
# nothing yet
system.time (makewinnerplots(showprogressbar=TRUE))
makewinnerplots(csvs = csvs[1:20])
}
}
