% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize aggregated data}
\usage{
normalize(aggregated_data)
}
\arguments{
\item{aggregated_data}{A dataframe representing data aggregated by Fast Track.}
}
\value{
Another dataframe with normalized formant frequencies.
}
\description{
Normalize aggregated data
}
\examples{
1+1
}
