% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogramplots.R
\name{plotffs}
\alias{plotffs}
\title{Plot an aggregated file}
\usage{
plotffs(
  ffs,
  xlim = NA,
  ylim = NA,
  xlab = NA,
  ylab = NA,
  main = NA,
  colors = NA,
  add = FALSE,
  spect = NA,
  ...
)
}
\arguments{
\item{ffs}{a dataframe containing formant tracks for a single file, or a list of dataframes comparing multiple analyses for a single file.}

\item{xlim}{an integer indicating which formant number should be plotted on the x axis.}

\item{ylim}{--.}

\item{xlab}{an optional user-specified x-axis label.}

\item{ylab}{an optional user-specified y-axis label.}

\item{main}{--.}

\item{colors}{--.}

\item{add}{if FALSE, a new plot if created.}

\item{spect}{a spectrogram to be shown behind the tracks.}

\item{...}{Additional arguments are passed to the internal call of 'plot'.}
}
\description{
Set up a plot aggregated files. This function does not draw anything but is used to set up the plot for other functions.
}
\examples{
\dontrun{
sound = readWave2("yoursound.wav")
ffs = analyze (sound)
spect = spectrogram (sound)
plotffs (ffs[[9]])
plotffs (ffs[[9]], spect = spect)
plotffs (ffs)
plotffs (ffs, spect = spect)
}
}
