% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{readcsvs}
\alias{readcsvs}
\title{Load Fast Track csv files}
\usage{
readcsvs(path = NA)
}
\arguments{
\item{path}{the path to the working directory for the Fast Track project.}
}
\description{
If no path is provided the working directory is assumed to be the Fast Track directory. This is recommended as it means that is means all reading/writing can be done without ever providing a path.
}
\examples{
\dontrun{
csvs = readcsvs ()
}
}
