% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{readformants}
\alias{readformants}
\title{Load Fast Track formant objects}
\usage{
readformants(path)
}
\arguments{
\item{path}{The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.}
}
\value{
A list of lists of dataframes. The "external" list is as long as number of files that were analyzed. For each "external" list element there are \emph{n} "internal" list elements for \emph{n} analysis steps. For example, \code{formant[[32]][[3]]} contains information regarding the 3rd analysis option for the 32nd file.
}
\description{
This function quickly reads in all the files contained in the "/formants" folder within a Fast Track directory. Since Fast Track exports one formant object per analysis option per token, thre are usually a large number of files stored in that directory. This function makes it easy to read all that information in at once. Note that it may take a bit of time for all the data to be read in if there are many tokens that were analyzed in that directory.
}
\examples{
\dontrun{
formants = readformants ()
}
}
