% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrids.R
\name{readtextgrid}
\alias{readtextgrid}
\title{Load textgrid information into R}
\usage{
readtextgrid(path)
}
\arguments{
\item{path}{the path to the Praat textgrid file you want to read.}
}
\value{
A list of dataframes, one containing the data from each interval tier in the textgrid. Each dataframe is named after the tier.
}
\description{
Create a list of dataframes from a textgrid. Each list element is a dataframe representing information from a different interval tier. Each dataframe contains the interval label, start time, end time, and duration (all in milliseconds).
}
\examples{
\dontrun{
path = "/textgrids/your-textgrid-file.TextGrid"
readtextgrid (path)
}
}
