% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{spectrogram}
\alias{spectrogram}
\title{Spectrogram}
\usage{
spectrogram(
  sound,
  maxformant = 5000,
  windowlength = 0.005,
  timestep = 0.001,
  dynamicrange = 60,
  resolution = 50,
  plot = TRUE,
  ...
)
}
\arguments{
\item{sound}{a numeric vector representing the sound to be analyzed.}

\item{maxformant}{the maximum analysis frequency (i.e., the Nyquist/2).}

\item{windowlength}{the windowlength specified in seconds.}

\item{timestep}{the analysis time step specified in seconds.}

\item{dynamicrange}{the dynamic range desired for the spectrogram, in decibels.}

\item{resolution}{the spectral resolution in Hertz.}

\item{plot}{if TRUE, a plot is created.}

\item{...}{Additional arguments are passed to the internal call of 'image'.}
}
\value{
A matrix representing a spectrogram.
}
\description{
Make a spectrogram of a sound.
}
\examples{
\dontrun{
sound = readWave2("yoursound.wav")
spect = spectrogram (sound, maxformant = 5000)
ffs = analyze (sound, timestep = 2)
plotffs (ffs[[2]], spect = spect)
}
}
