% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackformants.R
\name{trackformants}
\alias{trackformants}
\title{Track formants for a sound or folder of sounds}
\usage{
trackformants(
  path = NA,
  from = 4800,
  to = 6800,
  nsteps = 12,
  windowlength = 0.05,
  n_formants = 3,
  timestep = 0.002,
  fileinformation = NA,
  estimateduration = TRUE
)
}
\arguments{
\item{path}{a Wave object read in using the readWave function from the tuneR package, or a list containing a set of these to be analyzed. A list of paths to wave files on your hard drive can be provided instead using the path parameter.}

\item{from}{the lowest analysis frequency.}

\item{to}{the highest analysis frequency.}

\item{nsteps}{the number of steps between the lowest and highest analysis frequencies.}

\item{windowlength}{the windowlength specified in seconds.}

\item{n_formants}{--.}

\item{timestep}{the analysis time step specified in seconds.}

\item{fileinformation}{--.}

\item{estimateduration}{--.}
}
\value{
A list of lists of dataframes. The 'external' list is as long as number of files that were analyzed. For each 'external' list element there are N 'internal' list elements, for N analysis steps. For example, 'formant[[32]][[3]]' contains information regarding the 3rd analysis option for the 32nd file.
}
\description{
Track formants for a sound or folder of sounds
}
\examples{
\dontrun{
sound = readWave2("yoursound.wav")
ffs1 = analyze (sound, timestep = 0.002, from = 4800, to = 6800, nsteps=12)
ffs2 = trackformants (sound, timestep = 0.002)
plotffs (ffs1)
plotffs (ffs1[[2]])
}
}
