% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackformants.R
\name{trackformants}
\alias{trackformants}
\title{Track formants}
\usage{
trackformants(sound, maxformant = 5000, windowlength = 0.05, timestep = 0.0025)
}
\arguments{
\item{sound}{a numeric vector representing the sound to be analyzed.}

\item{maxformant}{the maximum analysis frequency (i.e., the Nyquist/2).}

\item{windowlength}{the windowlength specified in seconds.}

\item{timestep}{the analysis time step specified in seconds.}
}
\value{
A matrix in which each row represents a different time point. The first four columns represent the frequencies of F1-F4, and columns 5-8 represent their bandwidths.
}
\description{
Track formants
}
\examples{
\dontrun{
sound = readWave2("yoursound.wav")
trackformants (sound, maxformant = 5000)
}
}
